Massive refactor, working dragging of sprite
I basically realized I had seriously misunderstood how brute expects things to be. Now, decisions on when to run code is hard-coded in core.cljs, and entity-specific code has been moved to their own files.

Eventually, system functions will be registered to perform updates, and that is the type of division I wanted.